requirements.txt
requests
psutil
GitPython
click
click-help-colors
colorama;platform_system=="Windows"
